restructure of the system directory
Public: Header files which should be exposed (.h)

Private: Implementation of headerfiles (.cpp)
