+++ Source/NetworkingTestCPP/Private/AbilitySystem/AbilityGameplayActors/S_GroundTarget.cpp
@@ -1,3 +1,3 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
-#include "S_PlayerStats.h"
+#include "AbilitySystem/AbilityGameplayActors/S_GroundTarget.h"
