+++ Source/NetworkingTestCPP/Private/AbilitySystem/AbilityGameplayActors/S_SpellAOEInfo.cpp
@@ -1,0 +1,4 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "AbilitySystem/AbilityGameplayActors/S_SpellAOEInfo.h"
+
