+++ Source/NetworkingTestCPP/Private/AbilitySystem/TargetTypes.cpp
@@ -1,6 +1,6 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
-#include "TargetTypes.h"
+#include "AbilitySystem/TargetTypes.h"
 
 
 bool FTargetDataHandle::NetSerialize(FArchive& Ar, class UPackageMap* Map, bool& bOutSuccess)
