+++ Source/NetworkingTestCPP/Private/Data/E_DamageTypes.cpp
@@ -1,3 +1,4 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
-#include "S_ProjectileTargeting.h"
+#include "Data/E_DamageTypes.h"
+
