+++ Source/NetworkingTestCPP/Private/Data/E_ResourceTypes.cpp
@@ -1,3 +1,3 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
-#include "E_ResourceTypes.h"
+#include "Data/E_ResourceTypes.h"
