+++ Source/NetworkingTestCPP/Private/Player/S_PlayerStats.cpp
@@ -1,4 +1,3 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
-#include "S_SpellAOEInfo.h"
-
+#include "Player/S_PlayerStats.h"
